+++ CHANGELOG.rst
@@ -11,10 +11,14 @@ The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 
 Added
 +++++
-Changes to improve and change way fitting is done. Main conceptual difference is to first fit curves with fixed slope, then re-fit all parameters:
+- Changes to improve and change way fitting is done. Main conceptual difference is to first fit curves with fixed slope, then re-fit all parameters:
   - Add ``test_curves.ipynb`` example.
   - Add ``init_slope`` to ``HillCurve`` and ``CurveFits``
   - Add ``fix_slope_first`` to ``HillCurve`` and ``CurveFits`` and make it True by default.
+
+- Improve exception handling:
+  - Add ``HillCurveFittingError``, and when fitting fails raise this exception rather than ``RuntimeError`` as was done previously.
+  - Improve error messages for some exceptions and change a few exception types.
 
 0.10.1
 ------